third_party/WebKit/LayoutTests/css3/motion-path/use-count-motion-path.html

<!DOCTYPE html>
<head>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
  <style>
  .n { motion-path: none; }
  </style>
</head>

<div id="target"></div>
<script>
var CSSMotionInEffect = 1616; // From UseCounter.h
var CSSOffsetInEffect = 1617; // From UseCounter.h

test(function() {
  assert_false(internals.isUseCounted(document, CSSMotionInEffect));

  var styleElement = document.createElement('style');
  styleElement.textContent = ".p { motion-path: path('m 0 0 h 1'); }";
  document.head.appendChild(styleElement);

  assert_true(internals.isUseCounted(document, CSSMotionInEffect));

  assert_false(internals.isUseCounted(document, CSSOffsetInEffect));
});
</script>
